add unit tests for the web dropcap transformation
