+++ packages/article-skeleton/src/article-skeleton.web.js
@@ -78,13 +78,6 @@ class ArticleSkeleton extends Component {
       newContent = insertDropcapIntoAST(newContent, template, dropcapsDisabled);
     }
 
-    if (newContent.length > 0 && newContent[0].name === "dropCap") {
-      // remove the wrapping paragraph
-      newContent[0].children = newContent[0].children[0].children;
-      newContent[1].children = [newContent[0], ...newContent[1].children];
-      newContent = newContent.slice(1);
-    }
-
     receiveChildList([
       {
         elementId: "related-articles",
