+++ packages/article-skeleton/src/dropcap-util.web.js
@@ -1,0 +1,12 @@
+import insertDropcapIntoAST from "./dropcap-util-common";
+
+export default (children, template, isDropcapDisabled) => {
+  let newContent = insertDropcapIntoAST(children, template, isDropcapDisabled);
+  if (newContent.length > 0 && newContent[0].name === "dropCap") {
+    // remove the wrapping paragraph
+    newContent[0].children = newContent[0].children[0].children;
+    newContent[1].children = [newContent[0], ...newContent[1].children];
+    newContent = newContent.slice(1);
+  }
+  return newContent;
+};
